django_backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
pillow==11.2.1
psycopg==3.2.9
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
